backend/requirements.txt
fastapi
uvicorn
pydantic
openai
python-dotenv
resend
requests
beautifulsoup4
